+++ custom/src/AuterionPlugin.cc
@@ -115,22 +115,20 @@ QVariantList&
 AuterionPlugin::settingsPages()
 {
     if(_auterionSettingsList.isEmpty()) {
-        addSettingsEntry(tr("General"),
-                       "qrc:/qml/GeneralSettings.qml","qrc:/res/gear-white.svg");
-        addSettingsEntry(tr("Comm Links"),
-                       "qrc:/qml/LinkSettings.qml", "qrc:/res/waves.svg");
-        addSettingsEntry(tr("Offline Maps"),
-                       "qrc:/qml/OfflineMap.qml", "qrc:/res/waves.svg");
+        addSettingsEntry(tr("General"),     "qrc:/qml/GeneralSettings.qml", "qrc:/res/gear-white.svg");
+        addSettingsEntry(tr("Comm Links"),  "qrc:/qml/LinkSettings.qml",    "qrc:/res/waves.svg");
+        addSettingsEntry(tr("Offline Maps"),"qrc:/qml/OfflineMap.qml",      "qrc:/res/waves.svg");
+#if defined(QGC_GST_MICROHARD_ENABLED)
+        addSettingsEntry(tr("Microhard"),   "qrc:/qml/MicrohardSettings.qml");
+#endif
 #if defined(QGC_GST_TAISYNC_ENABLED)
-        addSettingsEntry(tr("Taisync"), "qrc:/qml/TaisyncSettings.qml");
+        addSettingsEntry(tr("Taisync"),     "qrc:/qml/TaisyncSettings.qml");
 #endif
 #if defined(QGC_AIRMAP_ENABLED)
         addSettingsEntry(tr("AirMap"), "qrc:/qml/AirmapSettings.qml");
 #endif
-        addSettingsEntry(tr("MAVLink"),
-                       "qrc:/qml/MavlinkSettings.qml", "qrc:/res/waves.svg");
-        addSettingsEntry(tr("Console"),
-                       "qrc:/qml/QGroundControl/Controls/AppMessages.qml");
+        addSettingsEntry(tr("MAVLink"),"qrc:/qml/MavlinkSettings.qml", "    qrc:/res/waves.svg");
+        addSettingsEntry(tr("Console"),"qrc:/qml/QGroundControl/Controls/AppMessages.qml");
 #if defined(QT_DEBUG)
         //-- These are always present on Debug builds
         addSettingsEntry(tr("Mock Link"), "qrc:/qml/MockLink.qml");