+++ src/FirmwarePlugin/PX4/PX4FirmwarePlugin.cc
@@ -576,16 +576,6 @@ bool PX4FirmwarePlugin::vehicleYawsToNextWaypointInMission(const Vehicle* vehicl
     return true;
 }
 
-QGCCameraManager* PX4FirmwarePlugin::createCameraManager(Vehicle* vehicle)
-{
-    return new QGCCameraManager(vehicle);
-}
-
-QGCCameraControl* PX4FirmwarePlugin::createCameraControl(const mavlink_camera_information_t* info, Vehicle *vehicle, int compID, QObject* parent)
-{
-    return new QGCCameraControl(info, vehicle, compID, parent);
-}
-
 uint32_t PX4FirmwarePlugin::highLatencyCustomModeTo32Bits(uint16_t hlCustomMode)
 {
     union px4_custom_mode px4_cm;